+++ Source/src/Layers/SFHUDLayer.h
@@ -15,9 +15,9 @@
 
 - (void)gameScoreChanged:(int)newGameScore;
 
-- (void)updateHealthBarWithHealthInPercent:(float)healthInPercent;
+- (void)updateHealthBarWithHealthInPercent:(double)healthInPercent;
 
-- (void)updateShieldBarWithShieldInPercent:(float)shieldInPercent;
+- (void)updateShieldBarWithShieldInPercent:(double)shieldInPercent;
 
 
 @end
